--- a/src/main/scala/com/johnsnowlabs/storage/StorageLocator.scala
+++ b/src/main/scala/com/johnsnowlabs/storage/StorageLocator.scala
@@ -43,7 +43,7 @@ case class StorageLocator(database: String, storageRef: String, sparkSession: Sp
   val clusterFilePath: Path = {
     if (!getTmpLocation.matches("s3[a]?:/.*")) {
       Path.mergePaths(
-        new Path(fileSystem.getUri.toString + clusterTmpLocation),
+        new Path(clusterTmpLocation),
         new Path("/" + clusterFileName))
     } else new Path(clusterTmpLocation + "/" + clusterFileName)
   }