@maziyarpanahi 2024-01-11
@danilojsl perhaps this part can have logics for different storage layers like s3, dbfs, hdfs, local, etc.?

```
val clusterFilePath: Path = {
     if (!getTmpLocation.matches("s3[a]?:/.*")) {
       Path.mergePaths(
         new Path(fileSystem.getUri.toString + clusterTmpLocation),
         new Path("/" + clusterFileName))
     } else new Path(clusterTmpLocation + "/" + clusterFileName)
   }
```

Could you please have a look and run some tests?