@jiamaozheng 2024-01-15
> @danilojsl perhaps this part can have logics for different storage layers like s3, dbfs, hdfs, local, etc.?
> 
> ```
> val clusterFilePath: Path = {
>      if (!getTmpLocation.matches("s3[a]?:/.*")) {
>        Path.mergePaths(
>          new Path(fileSystem.getUri.toString + clusterTmpLocation),
>          new Path("/" + clusterFileName))
>      } else new Path(clusterTmpLocation + "/" + clusterFileName)
>    }
> ```
> 
> Could you please have a look and run some tests?

@maziyarpanahi , I have done some tests in HDFS, DBFS, S3 and Local environments, and the outcomes of these runs are expected. @danilojsl, please feel free to run more tests in your environments if needed. Also please approve three appending CI builds. Thanks,