@jiamaozheng 2024-01-22
> @maziyarpanahi I also ran several tests and the change is working. Thanks for the contribution @jiamaozheng

@maziyarpanahi , if there are no other concerns, would you please approve, merge and release this bug fix? Thanks,