@maziyarpanahi 2024-01-24
> > @maziyarpanahi I also ran several tests and the change is working. Thanks for the contribution @jiamaozheng
> 
> @maziyarpanahi , if there are no other concerns, would you please approve, merge and release this bug fix? Thanks,

Thanks @jiamaozheng 
This will be merged into our next release candidates and will be included in 5.2.4 release